amp head tag for experimental ng-interactive pages
