+++ applications/app/pages/InteractiveHtmlPage.scala
@@ -16,7 +16,8 @@ import views.html.fragments.page.head.stylesheets.{criticalStyleInline, critical
 import views.html.fragments.page.{devTakeShot, htmlTag}
 import views.html.fragments._
 import views.html.stacked
-import html.HtmlPageHelpers.{ContentCSSFile}
+import html.HtmlPageHelpers.ContentCSSFile
+import services.ApplicationsSpecial2020Election
 
 object InteractiveHtmlPage extends HtmlPage[InteractivePage] {
 
@@ -50,9 +51,12 @@ object InteractiveHtmlPage extends HtmlPage[InteractivePage] {
       ),
     )
 
+    val ampTag = ApplicationsSpecial2020Election.ampTagHtml(request.path)
+
     htmlTag(
       headTag(
         weAreHiring() when WeAreHiring.isSwitchedOn,
+        ampTag,
         titleTag(),
         metaData(),
         styles(allStyles),
