+++ applications/app/services/ApplicationsDotcomRenderingInterface.scala
@@ -1,7 +1,7 @@
 package services
 
 import experiments.{ActiveExperiments, NGInteractiveDCR}
-import play.api.mvc.{RequestHeader}
+import play.api.mvc.RequestHeader
 
 sealed trait RenderingTier
 object Legacy extends RenderingTier
@@ -28,20 +28,4 @@ object ApplicationsDotcomRenderingInterface {
   def getHtmlFromDCR(): String = {
     "Experiment: NGInteractiveDCR (2)"
   }
-}
-
-object ApplicationsSpecial2020Election {
-  def atomIdToCapiPath(atomId: String): String = {
-    /*
-        This function transforms and atom id
-          "interactives/2020/07/interactive-vaccine-tracker/default"
-        into the corresponding amp capi query path
-          "atom/interactive/interactives/2020/07/interactive-vaccine-tracker/amp-page"
-
-        I realise that this code is a little bit fragile, hopefully we will double check beforehand
-        that the logic still applies for the election result tracker.
-     */
-    (Array("atom", "interactive") ++ atomId.split("/").dropRight(1) ++ Array("amp-page")).mkString("/")
-  }
-
 }
